pyproject.toml
[tool.poetry]
name = "portkatz"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
pycryptodome = "^3.9.7"
psutil = "^5.7.0"
tqdm = "^4.46.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
